Move HazeProgressive to shared core
