+++ haze-blur/api/api.txt
@@ -11,7 +11,7 @@ package dev.chrisbanes.haze.blur {
     method @InaccessibleFromKotlin public dev.chrisbanes.haze.blur.HazeColorEffect getFallbackTint();
     method @InaccessibleFromKotlin public androidx.compose.ui.graphics.Brush? getMask();
     method @InaccessibleFromKotlin public float getNoiseFactor();
-    method @InaccessibleFromKotlin public dev.chrisbanes.haze.blur.HazeProgressive? getProgressive();
+    method @InaccessibleFromKotlin public dev.chrisbanes.haze.HazeProgressive? getProgressive();
     method @InaccessibleFromKotlin public dev.chrisbanes.haze.blur.HazeBlurStyle getStyle();
     method @InaccessibleFromKotlin public void setAlpha(float);
     method @InaccessibleFromKotlin public void setBlurEnabled(boolean);
@@ -19,7 +19,7 @@ package dev.chrisbanes.haze.blur {
     method @InaccessibleFromKotlin public void setFallbackTint(dev.chrisbanes.haze.blur.HazeColorEffect);
     method @InaccessibleFromKotlin public void setMask(androidx.compose.ui.graphics.Brush?);
     method @InaccessibleFromKotlin public void setNoiseFactor(float);
-    method @InaccessibleFromKotlin public void setProgressive(dev.chrisbanes.haze.blur.HazeProgressive?);
+    method @InaccessibleFromKotlin public void setProgressive(dev.chrisbanes.haze.HazeProgressive?);
     method @InaccessibleFromKotlin public void setStyle(dev.chrisbanes.haze.blur.HazeBlurStyle);
     property public float alpha;
     property public androidx.compose.ui.graphics.Color backgroundColor;
@@ -30,7 +30,7 @@ package dev.chrisbanes.haze.blur {
     property public dev.chrisbanes.haze.blur.HazeColorEffect fallbackTint;
     property public androidx.compose.ui.graphics.Brush? mask;
     property public float noiseFactor;
-    property public dev.chrisbanes.haze.blur.HazeProgressive? progressive;
+    property public dev.chrisbanes.haze.HazeProgressive? progressive;
     property public dev.chrisbanes.haze.blur.HazeBlurStyle style;
     field public static final String TAG = "BlurVisualEffect";
   }
@@ -139,57 +139,6 @@ package dev.chrisbanes.haze.blur {
     method public static inline void blurEffect(dev.chrisbanes.haze.HazeEffectScope, kotlin.jvm.functions.Function1<? super dev.chrisbanes.haze.blur.BlurVisualEffect,kotlin.Unit> block);
   }
 
-  @androidx.compose.runtime.Immutable public sealed exhaustive interface HazeProgressive {
-    field public static final dev.chrisbanes.haze.blur.HazeProgressive.Companion Companion;
-  }
-
-  @kotlin.jvm.JvmInline public static final value class HazeProgressive.Brush implements dev.chrisbanes.haze.blur.HazeProgressive {
-    ctor @KotlinOnly public HazeProgressive.Brush(androidx.compose.ui.graphics.Brush brush);
-    method @InaccessibleFromKotlin public androidx.compose.ui.graphics.Brush getBrush();
-    property public androidx.compose.ui.graphics.Brush brush;
-  }
-
-  public static final class HazeProgressive.Companion {
-    method @KotlinOnly public inline dev.chrisbanes.haze.blur.HazeProgressive.Brush forShader(kotlin.jvm.functions.Function1<androidx.compose.ui.geometry.Size,androidx.compose.ui.graphics.Shader> block);
-    method public dev.chrisbanes.haze.blur.HazeProgressive.LinearGradient horizontalGradient(optional androidx.compose.animation.core.Easing easing, optional float startX, optional float startIntensity, optional float endX, optional float endIntensity, optional boolean preferPerformance);
-    method public dev.chrisbanes.haze.blur.HazeProgressive.LinearGradient verticalGradient(optional androidx.compose.animation.core.Easing easing, optional float startY, optional float startIntensity, optional float endY, optional float endIntensity, optional boolean preferPerformance);
-  }
-
-  public static final class HazeProgressive.LinearGradient implements dev.chrisbanes.haze.blur.HazeProgressive {
-    ctor public HazeProgressive.LinearGradient();
-    ctor @KotlinOnly public HazeProgressive.LinearGradient(optional androidx.compose.animation.core.Easing easing, optional androidx.compose.ui.geometry.Offset start, optional float startIntensity, optional androidx.compose.ui.geometry.Offset end, optional float endIntensity, optional boolean preferPerformance);
-    method public androidx.compose.animation.core.Easing component1();
-    method @KotlinOnly public operator androidx.compose.ui.geometry.Offset component2();
-    method public float component3();
-    method @KotlinOnly public operator androidx.compose.ui.geometry.Offset component4();
-    method public float component5();
-    method public boolean component6();
-    method @KotlinOnly public dev.chrisbanes.haze.blur.HazeProgressive.LinearGradient copy(optional androidx.compose.animation.core.Easing easing, optional androidx.compose.ui.geometry.Offset start, optional float startIntensity, optional androidx.compose.ui.geometry.Offset end, optional float endIntensity, optional boolean preferPerformance);
-    method @InaccessibleFromKotlin public androidx.compose.animation.core.Easing getEasing();
-    method @InaccessibleFromKotlin public float getEndIntensity();
-    method @InaccessibleFromKotlin public boolean getPreferPerformance();
-    method @InaccessibleFromKotlin public float getStartIntensity();
-    property public androidx.compose.animation.core.Easing easing;
-    property public androidx.compose.ui.geometry.Offset end;
-    property public float endIntensity;
-    property public boolean preferPerformance;
-    property public androidx.compose.ui.geometry.Offset start;
-    property public float startIntensity;
-  }
-
-  @dev.chrisbanes.haze.Poko public static final class HazeProgressive.RadialGradient implements dev.chrisbanes.haze.blur.HazeProgressive {
-    ctor public HazeProgressive.RadialGradient();
-    ctor @KotlinOnly public HazeProgressive.RadialGradient(optional androidx.compose.animation.core.Easing easing, optional androidx.compose.ui.geometry.Offset center, optional float centerIntensity, optional float radius, optional float radiusIntensity);
-    method @InaccessibleFromKotlin public float getCenterIntensity();
-    method @InaccessibleFromKotlin public androidx.compose.animation.core.Easing getEasing();
-    method @InaccessibleFromKotlin public float getRadius();
-    method @InaccessibleFromKotlin public float getRadiusIntensity();
-    property public androidx.compose.ui.geometry.Offset center;
-    property public float centerIntensity;
-    property public androidx.compose.animation.core.Easing easing;
-    property public float radius;
-    property public float radiusIntensity;
-  }
+  @Deprecated public typealias HazeProgressive = dev.chrisbanes.haze.HazeProgressive;
 
 }
-
