+++ haze/api/api.txt
@@ -145,6 +145,62 @@ package dev.chrisbanes.haze {
     enum_constant public static final dev.chrisbanes.haze.HazePositionStrategy Screen;
   }
 
+  @androidx.compose.runtime.Immutable public sealed exhaustive interface HazeProgressive {
+    field public static final dev.chrisbanes.haze.HazeProgressive.Companion Companion;
+  }
+
+  @kotlin.jvm.JvmInline public static final value class HazeProgressive.Brush implements dev.chrisbanes.haze.HazeProgressive {
+    ctor @KotlinOnly public HazeProgressive.Brush(androidx.compose.ui.graphics.Brush brush);
+    method @InaccessibleFromKotlin public androidx.compose.ui.graphics.Brush getBrush();
+    property public androidx.compose.ui.graphics.Brush brush;
+  }
+
+  public static final class HazeProgressive.Companion {
+    method @KotlinOnly public inline dev.chrisbanes.haze.HazeProgressive.Brush forShader(kotlin.jvm.functions.Function1<androidx.compose.ui.geometry.Size,androidx.compose.ui.graphics.Shader> block);
+    method public dev.chrisbanes.haze.HazeProgressive.LinearGradient horizontalGradient(optional androidx.compose.animation.core.Easing easing, optional float startX, optional float startIntensity, optional float endX, optional float endIntensity, optional boolean preferPerformance);
+    method public dev.chrisbanes.haze.HazeProgressive.LinearGradient verticalGradient(optional androidx.compose.animation.core.Easing easing, optional float startY, optional float startIntensity, optional float endY, optional float endIntensity, optional boolean preferPerformance);
+  }
+
+  public static final class HazeProgressive.LinearGradient implements dev.chrisbanes.haze.HazeProgressive {
+    ctor public HazeProgressive.LinearGradient();
+    ctor @KotlinOnly public HazeProgressive.LinearGradient(optional androidx.compose.animation.core.Easing easing, optional androidx.compose.ui.geometry.Offset start, optional float startIntensity, optional androidx.compose.ui.geometry.Offset end, optional float endIntensity, optional boolean preferPerformance);
+    method public androidx.compose.animation.core.Easing component1();
+    method @KotlinOnly public operator androidx.compose.ui.geometry.Offset component2();
+    method public float component3();
+    method @KotlinOnly public operator androidx.compose.ui.geometry.Offset component4();
+    method public float component5();
+    method public boolean component6();
+    method @KotlinOnly public dev.chrisbanes.haze.HazeProgressive.LinearGradient copy(optional androidx.compose.animation.core.Easing easing, optional androidx.compose.ui.geometry.Offset start, optional float startIntensity, optional androidx.compose.ui.geometry.Offset end, optional float endIntensity, optional boolean preferPerformance);
+    method @InaccessibleFromKotlin public androidx.compose.animation.core.Easing getEasing();
+    method @InaccessibleFromKotlin public float getEndIntensity();
+    method @InaccessibleFromKotlin public boolean getPreferPerformance();
+    method @InaccessibleFromKotlin public float getStartIntensity();
+    property public androidx.compose.animation.core.Easing easing;
+    property public androidx.compose.ui.geometry.Offset end;
+    property public float endIntensity;
+    property public boolean preferPerformance;
+    property public androidx.compose.ui.geometry.Offset start;
+    property public float startIntensity;
+  }
+
+  @dev.chrisbanes.haze.Poko public static final class HazeProgressive.RadialGradient implements dev.chrisbanes.haze.HazeProgressive {
+    ctor public HazeProgressive.RadialGradient();
+    ctor @KotlinOnly public HazeProgressive.RadialGradient(optional androidx.compose.animation.core.Easing easing, optional androidx.compose.ui.geometry.Offset center, optional float centerIntensity, optional float radius, optional float radiusIntensity);
+    method @InaccessibleFromKotlin public float getCenterIntensity();
+    method @InaccessibleFromKotlin public androidx.compose.animation.core.Easing getEasing();
+    method @InaccessibleFromKotlin public float getRadius();
+    method @InaccessibleFromKotlin public float getRadiusIntensity();
+    property public androidx.compose.ui.geometry.Offset center;
+    property public float centerIntensity;
+    property public androidx.compose.animation.core.Easing easing;
+    property public float radius;
+    property public float radiusIntensity;
+  }
+
+  public final class HazeProgressiveKt {
+    method @dev.chrisbanes.haze.InternalHazeApi public static androidx.compose.ui.graphics.Brush asBrush(dev.chrisbanes.haze.HazeProgressive, optional int numStops);
+  }
+
   @dev.chrisbanes.haze.ExperimentalHazeApi public final class HazeSourceNode extends androidx.compose.ui.Modifier.Node implements androidx.compose.ui.node.CompositionLocalConsumerModifierNode androidx.compose.ui.node.DrawModifierNode androidx.compose.ui.node.GlobalPositionAwareModifierNode androidx.compose.ui.node.LayoutAwareModifierNode androidx.compose.ui.node.ObserverModifierNode androidx.compose.ui.node.TraversableNode {
     ctor public HazeSourceNode(dev.chrisbanes.haze.HazeState state, optional float zIndex, optional Object? key);
     method public void draw(androidx.compose.ui.graphics.drawscope.ContentDrawScope);
@@ -237,4 +293,3 @@ package dev.chrisbanes.haze {
   }
 
 }
-
